move stuff around; fix names
