+++ web/src/components/App.js
@@ -15,7 +15,7 @@ import StateContacts from '../containers/StateContacts';
 import StateStart from '../containers/StateStart';
 import StatePersonnel from '../containers/StatePersonnel';
 import ActivitySchedule from '../containers/ActivitySchedule';
-import Expenses from '../containers/Expenses';
+import ActivityExpenses from '../containers/ActivityExpenses';
 
 const Wrapper = process.env.NODE_ENV !== 'production' ? DevProvider : Provider;
 
@@ -40,7 +40,7 @@ const App = () => (
         <Route path="/activity-overview" component={ActivityOverview} />
         <Route path="/activity-schedule" component={ActivitySchedule} />
         <Route path="/apd-overview" component={ApdOverview} />
-        <Route path="/expenses" component={Expenses} />
+        <Route path="/expenses" component={ActivityExpenses} />
         <Route path="/state-contacts" component={StateContacts} />
         <Route path="/state-personnel" component={StatePersonnel} />
         <Route path="/state-start" component={StateStart} />
