--- web/src/components/Form/Expenses/index.js
@@ -1,9 +1,0 @@
-import CommonExpenses from './CommonExpenses';
-import ExpenseCategory from './ExpenseCategory';
-
-export default {
-  CommonExpenses,
-  ExpenseCategory
-};
-
-export { CommonExpenses, ExpenseCategory };
